myservies: Remove 99-sound
Before 99-sound used to fail if the custom playback device wasn't
loaded. However, it won't be loaded until audio starts playing. Now it
just issues the command to reload state and exits.

It was going to be made to just issue the reload but at the point may as
well just add it to rc.local