linux: Make linux-headers useful again
linux-headers and linux use the same source. There are already
workarounds to avoid download the source twice, but linux grew to just
consume linux-headers.

While this generally works, it seems annoying to deal with when packages
just want the headers to build. So now we split the headers back into
their own package. This removes the need of the "provides" directory

Also fix a missing build depedency of bison for linux